FEATURE: Add Connections to README page
Added a 'Logic App Connections' section to the start page.
This lists the global connections for the standard logic app.  If the script is downloading
from Azure, then additional look up functions are completed to get additional
information about some of the connections.  This enhanced information is written back to the
working directory to enable the script to be run in offline mode.

Changed Start.md to README.md to align to with Git repositories